move nodeExecution types to job
