+++ ui/src/types/index.ts
@@ -7,7 +7,6 @@ export * from "./layout";
 export * from "./log";
 export * from "./member";
 export * from "./node";
-export * from "./nodeExecutions";
 export * from "./project";
 export * from "./shared";
 export * from "./shortcuts";
