+++ ui/src/types/job.ts
@@ -1,5 +1,14 @@
 import type { ApiResponse } from "./api";
-import type { NodeExecution } from "./nodeExecutions";
+
+export type NodeStatus = "pending" | "running" | "succeeded" | "failed";
+
+export type NodeExecution = {
+  nodeId: string;
+  status: NodeStatus;
+  startedAt?: string;
+  completedAt?: string;
+  intermediateDataUrl?: string;
+};
 
 export type JobStatus =
   | "queued"
