--- ui/src/types/nodeExecutions.ts
@@ -1,9 +1,0 @@
-export type NodeStatus = "pending" | "running" | "succeeded" | "failed";
-
-export type NodeExecution = {
-  nodeId: string;
-  status: NodeStatus;
-  startedAt?: string;
-  completedAt?: string;
-  intermediateDataUrl?: string;
-};
